Correct allele query order with original allele order
Refactor
